fix #16: fix code style
