+++ src/main/java/cz/cvut/iarylser/dao/dto/JwtAuthenticationResponse.java
@@ -1,13 +1,14 @@
-package cz.cvut.iarylser.dao.DTO;
+package cz.cvut.iarylser.dao.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class LikeRequest {
-    private Long eventId;
-    private Long userId;
+public class JwtAuthenticationResponse {
+    private String token;
 }
