+++ src/main/java/cz/cvut/iarylser/dao/dto/LikeRequest.java
@@ -1,14 +1,13 @@
-package cz.cvut.iarylser.dao.DTO;
+package cz.cvut.iarylser.dao.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class JwtAuthenticationResponse {
-    private String token;
+public class LikeRequest {
+    private Long eventId;
+    private Long userId;
 }
