+++ src/main/java/cz/cvut/iarylser/dao/dto/SignInRequest.java
@@ -1,4 +1,4 @@
-package cz.cvut.iarylser.dao.DTO;
+package cz.cvut.iarylser.dao.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SignUpRequest {
+public class SignInRequest {
     private String username;
     private String password;
-    private String email;
-    private int age;
 }
