+++ src/main/java/cz/cvut/iarylser/dao/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package cz.cvut.iarylser.dao.DTO;
+package cz.cvut.iarylser.dao.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class SignInRequest {
+public class SignUpRequest {
     private String username;
     private String password;
+    private String email;
+    private int age;
 }
