+++ src/main/java/cz/cvut/iarylser/dao/mappersDto/MapperDTO.java
@@ -1,4 +1,4 @@
-package cz.cvut.iarylser.dao.mappersDTO;
+package cz.cvut.iarylser.dao.mappersDto;
 
 import java.util.List;
