+++ src/main/java/cz/cvut/iarylser/dao/mappersDto/TicketMapperDTO.java
@@ -1,6 +1,6 @@
-package cz.cvut.iarylser.dao.mappersDTO;
+package cz.cvut.iarylser.dao.mappersDto;
 
-import cz.cvut.iarylser.dao.DTO.TicketDTO;
+import cz.cvut.iarylser.dao.dto.TicketDTO;
 import cz.cvut.iarylser.dao.entity.Ticket;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
