+++ src/main/java/cz/cvut/iarylser/facade/TicketFacade.java
@@ -1,6 +1,6 @@
 package cz.cvut.iarylser.facade;
 
-import cz.cvut.iarylser.dao.DTO.TicketDTO;
+import cz.cvut.iarylser.dao.dto.TicketDTO;
 
 import java.util.List;
