+++ src/test/java/cz/cvut/iarylser/unit/facade/EventFacadeImplTest.java
@@ -2,8 +2,6 @@ package cz.cvut.iarylser.unit.facade;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 class EventFacadeImplTest {
 
     @Test
